NDWH-2811: Rename "dags" folder to be distinct after installation
