+++ src/maintenance_dags/__init__.py
@@ -1,0 +1,4 @@
+from .airflow_clear_missing_dags import clear_missing_dags_dag
+from .airflow_db_cleanup import db_cleanup_dag
+from .airflow_kill_halted_tasks import kill_halted_tasks_dag
+from .airflow_log_cleanup import log_cleanup_dag
